requirements.txt
slack-sdk==3.21.3
flask==2.2.3
slackeventsapi==3.0.1
waitress==2.1.2
python-dotenv==1.0.0
pandas==1.4.3
paramiko==2.11.0
PyPDF2>=2.10.9
pypdf ==3.12.0
gunicorn==20.1.0
requests==2.27.1
werkzeug==2.2.2

